+++ docker-compose-base.yaml
@@ -24,6 +24,7 @@ services:
       interval: 10s
       timeout: 5s
       retries: 3
+    command: --quiet
   order-dapr:
     image: "daprio/daprd:edge"
     command:
