+++ rustfmt.toml
@@ -1,0 +1,1 @@
+reorder_imports = true
