+++ src/event/model/mod.rs
@@ -1,0 +1,3 @@
+pub mod order_compensation_dto;
+pub mod order_dto;
+pub mod order_item_dto;
