+++ src/graphql/mod.rs
@@ -1,0 +1,4 @@
+pub mod model;
+pub mod mutation;
+pub mod mutation_input_structs;
+pub mod query;
