+++ src/graphql/model/connection/base_connection.rs
@@ -1,6 +1,7 @@
 use async_graphql::{OutputType, SimpleObject};
+use mongodb_cursor_pagination::FindResult;
 
-/// A base connection for an OutputType.
+/// A base connection for an output type.
 #[derive(SimpleObject)]
 #[graphql(shareable)]
 pub struct BaseConnection<T: OutputType> {
@@ -12,8 +13,6 @@ pub struct BaseConnection<T: OutputType> {
     pub total_count: u64,
 }
 
-use mongodb_cursor_pagination::FindResult;
-
 pub struct FindResultWrapper<Node>(pub FindResult<Node>);
 
 /// Object that writes total count of items in a query, regardless of pagination.
@@ -22,7 +21,7 @@ pub struct AdditionalFields {
     total_count: u64,
 }
 
-/// Implementation of conversion from MongoDB pagination to GraphQL Connection.
+/// Implementation of conversion from MongoDB pagination to GraphQL connection.
 impl<Node> From<FindResultWrapper<Node>> for BaseConnection<Node>
 where
     Node: OutputType,
