+++ src/graphql/model/connection/mod.rs
@@ -1,0 +1,5 @@
+pub mod base_connection;
+pub mod discount_connection;
+pub mod order_connection;
+pub mod order_item_connection;
+pub mod product_variant_version_connection;
