+++ src/graphql/model/mod.rs
@@ -1,0 +1,7 @@
+pub mod connection;
+pub mod foreign_types;
+pub mod order;
+pub mod order_datatypes;
+pub mod order_item;
+pub mod payment_authorization;
+pub mod user;
