+++ src/graphql/model/order.rs
@@ -1,36 +1,33 @@
 use std::cmp::Ordering;
 
-use async_graphql::{ComplexObject, Enum, Error, Result, SimpleObject};
+use async_graphql::{ComplexObject, Enum, Result, SimpleObject};
 use bson::Uuid;
 use bson::{datetime::DateTime, Bson};
 use serde::{Deserialize, Serialize};
 
-use crate::foreign_types::UserAddress;
-use crate::order_datatypes::OrderDirection;
-use crate::order_item::OrderItemDTO;
-use crate::payment_authorization::PaymentAuthorization;
-use crate::{
-    order_datatypes::CommonOrderInput, order_item::OrderItem,
-    order_item_connection::OrderItemConnection, user::User,
-};
+use super::connection::order_item_connection::OrderItemConnection;
+use super::foreign_types::UserAddress;
+use super::order_datatypes::{CommonOrderInput, OrderDirection};
+use super::order_item::OrderItem;
+use super::user::User;
 
-/// The Order of a user.
+/// The order of a user.
 #[derive(Debug, Serialize, Deserialize, PartialEq, Clone, SimpleObject)]
 #[graphql(complex)]
 pub struct Order {
     /// Order UUID.
     pub _id: Uuid,
-    /// User.
+    /// User owning order.
     pub user: User,
-    /// Timestamp when Order was created.
+    /// Timestamp when order was created.
     pub created_at: DateTime,
-    /// The status of the Order.
+    /// The status of the order.
     pub order_status: OrderStatus,
-    /// Timestamp of Order placement. `None` until Order is placed.
+    /// Timestamp of order placement. `None` until order is placed.
     pub placed_at: Option<DateTime>,
-    /// The rejection reason if status of the Order is `OrderStatus::Rejected`.
+    /// The rejection reason if status of the order is `OrderStatus::Rejected`.
     pub rejection_reason: Option<RejectionReason>,
-    /// The internal vector consisting of OrderItems.
+    /// The internal vector consisting of order items.
     #[graphql(skip)]
     pub internal_order_items: Vec<OrderItem>,
     /// Address to where the order should be shipped to.
@@ -79,7 +76,7 @@ impl Order {
     }
 }
 
-/// Describes if Order is placed, or yet pending. An Order can be rejected during its lifetime.
+/// Describes if order is placed, or yet pending. An order can be rejected during its lifetime.
 #[derive(Debug, Enum, Copy, Clone, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "SCREAMING_SNAKE_CASE")]
 pub enum OrderStatus {
@@ -107,7 +104,7 @@ impl From<OrderStatus> for Bson {
     }
 }
 
-/// Describes the reason why an Order was rejected, in case of rejection: `OrderStatus::Rejected`.
+/// Describes the reason why an order was rejected, in case of rejection: `OrderStatus::Rejected`.
 #[derive(Debug, Enum, Copy, Clone, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "SCREAMING_SNAKE_CASE")]
 pub enum RejectionReason {
@@ -137,69 +134,4 @@ fn sort_order_items(order_items: &mut Vec<OrderItem>, order_by: Option<CommonOrd
         true => Ordering::Less,
         false => Ordering::Greater,
     });
-}
-
-/// DTO of an order of a user.
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct OrderDTO {
-    /// Order UUID.
-    pub id: Uuid,
-    /// UUID of user connected with Order.
-    pub user_id: Uuid,
-    /// Timestamp when Order was created.
-    pub created_at: chrono::DateTime<chrono::Utc>,
-    /// The status of the Order.
-    pub order_status: OrderStatus,
-    /// Timestamp of Order placement. `None` until Order is placed.
-    pub placed_at: chrono::DateTime<chrono::Utc>,
-    /// The rejection reason if status of the Order is `OrderStatus::Rejected`.
-    pub rejection_reason: Option<RejectionReason>,
-    /// OrderItems associated with the order.
-    pub order_items: Vec<OrderItemDTO>,
-    /// UUID of address to where the order should be shipped to.
-    pub shipment_address_id: Uuid,
-    /// UUID of address of invoice.
-    pub invoice_address_id: Uuid,
-    /// Total compensatable amount of order.
-    pub compensatable_order_amount: u64,
-    /// UUID of payment information that the order should be processed with.
-    pub payment_information_id: Uuid,
-    /// Optional payment authorization information.
-    pub payment_authorization: Option<PaymentAuthorization>,
-    /// VAT number.
-    pub vat_number: String,
-}
-
-impl TryFrom<(Order, Option<PaymentAuthorization>)> for OrderDTO {
-    type Error = Error;
-
-    fn try_from(
-        (order, payment_authorization): (Order, Option<PaymentAuthorization>),
-    ) -> Result<Self, Self::Error> {
-        let order_item_dtos = order
-            .internal_order_items
-            .iter()
-            .map(|o| OrderItemDTO::from(o.clone()))
-            .collect();
-        let message =
-            format!("OrderDTO cannot be created, `placed_at` of the given Order is `None`");
-        let placed_at = order.placed_at.ok_or(Error::new(message))?.to_chrono();
-        let order_dto = Self {
-            id: order._id,
-            user_id: order.user._id,
-            created_at: order.created_at.to_chrono(),
-            order_status: order.order_status,
-            placed_at,
-            rejection_reason: order.rejection_reason,
-            order_items: order_item_dtos,
-            shipment_address_id: order.shipment_address._id,
-            invoice_address_id: order.invoice_address._id,
-            compensatable_order_amount: order.compensatable_order_amount,
-            payment_information_id: order.payment_information_id,
-            payment_authorization: payment_authorization,
-            vat_number: order.vat_number,
-        };
-        Ok(order_dto)
-    }
 }
