+++ src/graphql/mutation_input_structs.rs
